+++ nope/lib/nope/Nope/Platform/Setting/Field/Input.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Nope\Platform\Setting\Field;
+
+class Input extends AbstractField {
+
+  function draw() {
+    return '<input type="text" '. $this->getAttributesList() .' />';
+  }
+
+}