+++ nope/lib/nope/Nope/Platform/Setting/Field/Text.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Nope\Platform\Setting\Field;
+
+class Text extends AbstractField {
+
+  function draw() {
+    return '<textarea '. $this->getAttributesList() .' ></textarea>';
+  }
+
+}